renamed helpers.js to article-helpers.js
